🧹 Move logic to service
