+++ src/controllers/text.js
@@ -21,12 +21,8 @@ async function helpCommand(ctx) {
 }
 
 async function actionCommand(ctx) {
-    // FIXME: Move logic to service
-    // Service should return final string
-    const actionMessage = await action(ctx.rawData)
-    const name = ctx.from.first_name
-    const message = `${name}: *${actionMessage}*`
-    await ctx.text(message)
+    const actionMessage = await action(ctx.rawData, ctx.from.first_name)
+    await ctx.text(actionMessage)
     try {
         await ctx.deleteMessage()
     } catch {
@@ -75,7 +71,6 @@ async function broadcastCommand(ctx) {
             )
             sent++
         } catch ({ message }) {
-            // FIXME: Move error message to some storage
             console.info(`Can't broadcast message: ${message}`)
             // Never mind if I can't send it
         }
@@ -93,7 +88,6 @@ async function anonymousMessageCommand(ctx) {
         await ctx.deleteMessage()
         await ctx.text(messageText)
     } catch ({ message }) {
-        // FIXME: Move error message to some storage
         console.info(`Can't send anonymous message: ${message}`)
     }
 }
