+++ src/services/handlers/text/action.js
@@ -1,13 +1,6 @@
-// FIXME: Move replacements to some storage
-const replacements = [
-    ['йти', 'шел(-ла)'],
-    ['реть', 'р(-ла)'],
-    ['ться', 'лся(-ась)'],
-    ['хнуть', 'х(-ла)'],
-    ['ть', 'л(-а)'],
-]
+import { replacements } from '../../../static/replacements.js'
 
-async function action(text) {
+async function action(text, name) {
     let result = text
     const ending = `(?=[^А-ЯЁа-яё*]|$)`
     for (const [from, to] of replacements) {
@@ -15,7 +8,7 @@ async function action(text) {
         result = result.replace(regexp, to)
     }
     result = result.replace(/(?<=[А-ЯЁа-яё])\*(?=\s|\n|$)/g, '')
-    return result
+    return `${name}: *${result}*`
 }
