+++ src/static/replacements.js
@@ -1,0 +1,11 @@
+const replacements = [
+    ['йти', 'шел(-ла)'],
+    ['реть', 'р(-ла)'],
+    ['ться', 'лся(-ась)'],
+    ['хнуть', 'х(-ла)'],
+    ['ть', 'л(-а)'],
+]
+
+export {
+    replacements
+}
